env variable stopgap that works in dev and release
astro is really bad/inconsistent at handling this. i thought,
process.env would fix it, but no. now dev
is broken.

there is already an issue for getting this fixed: withastro/astro#9827
i hope they can deal with this in a somewhat timely manner, so we don't
need to rely on such hacks anymore