kinks: remove Children
This change has been long overdue. This "kink" was only introduced
because of KCT's ancestry (Fetish Chart 2014) and there is one important
characteristic that separates it from the other kinks: It cannot be
practiced consensually. This fact makes it unfit for any practical kink
discussion, rating it well would be nothing but a self-report.

For the reasons stated above, this change will also apply retroactively,
once we host old KCC versions again.